Add an option to map mseg pages with THP
In order for mseg pages to be reliably mapped with pages lager than
the default page size, the mapping must start and end at a multiple of
the larger page size.

To do this, this change adds an abstraction for performing a
memory-mapping with a specified alignment.  On an operating systems
like SunOS 5.9 and later, this is done by passing some extra flags to
mmap(2).  On operating systems without such a capability, we must do
this manually by over-allocating and freeing the excess.

The logic in this change only affects super carrier allocations but it
can be generalized to other mseg allocations.