Make application:stop/1 callback optional
The majority of the times it simply returns ok.
In Erlang/OTP itself, only ssl and snmp implemented it.

Note we didn't have to change the implementation
because the callback is always invoked as
`catch App:stop(Arg)`, which is also how other optional
callbacks, such as `prep_stop/1` are handled.