stdlib: Include process label in gen crashes
For gen_statem processes, as well as gen_servers, gen_event and
gen_fsms, if they crash and a process label had been set with
proc_lib:set_label/1, then include that label in the printout,
to make it easier to identify the crashing process.

Below is an example of what it may look like for a gen_statem.
The "Process label" on the second line is new.

  ** State machine <0.1386.0> terminating
  ** Process label = {some_label, ...}
  ** When server state  = #{...}
  ** Reason for termination = error:...
  ** Callback modules = [...]
  ** Callback mode = state_functions

If no label has been set, then that process label line will not be
included, and the printout will look like it did before this commit.