Check for embedded mode in the client
This ensures that checking for code:ensure_loaded/1
in embedded mode (and other operations) do not send
additional messages to the code server.

This pull request fully closes #7889. While previous
work has fixed the slow path in the code server, this
pull request makes it so the code server is avoided
altogether.