Optimize application:get_key to avoid full record copies
`application:get_key` would previously copy from ETS the entire appl
record and then select a single field from it. This can be optimized
to only select the specific field that's required with `ets:match`.

Especially for `application:get_key(X, env)` this would copy the
appl record, completely discard it and lookup the environment
which is stored separately. We still need to query separately
to preserve the behaviour of returning `undefined` for unloaded
applications.