Optimise ETS usage in code_server
Since the code_server state already holds a reference to the code_server
table that is meticulously passed around, we can look up the direct
reference with `ets:whereis`, rather than passing the name around.
ETS operations with a refrence are faster than with the name.

Additionally, for the `code_names` table, we can make it private
and unnamed completely, since it's only used directly from the server,
there's no access to it from outside the process.