Use code_server path cache in code:where_is_file/1
`application:load/1` slows down as the number of directories in the code
path increases because the call to `code:where_is_file/1` for the '.app'
file must scan each directory for the app. The code_server maintains
a cache of the contents of directories in the path depending on their
`code:cache()` setting. Re-using and updating that cache when searching
for '.app' files in `application:load/1` can improve its runtime,
especially when loading multiple applications.

For a reasonably large project (RabbitMQ) with a high number of code
path dirs (111 at time of writing), loading an application might take
upwards of 20ms. With the code_server's cache, loading each application
takes at most around half of a millisecond instead. `application:load/1`
is used for loading plugins in Rabbit's CLI for example, and this change
can save as much as 750ms for a run of `rabbitmqctl --help` (30% of the
total run time). It also improves the boot time for the broker since
`application:load/1` is used by `application:start/2`.