[erts] Fix dirty signal handling
Commit (1) fixed so that signals were handled by dirty signal handlers also
for processes scheduled for dirty execution, but not yet selected for dirty
execution. This, however, opened up for signal handling that could release the
main lock of the process prior to dirty execution that inspected the signal
queue which meant that the dirty scheduler could see an inconsistent signal
queue with a crash as a result. This commit fixes so that also dirty
schedulers that may inspect the signal queue will release the main lock and
wait for signal handling to complete before inspecting the signal queue
when necessary.

(1) 00089cf03a40cc0b588e36949cf31d311d1e883f