wx: Add define to silence macro deprecation warnings
Since we run -Werror in some builds, we need to be able to silence
these types of warnings, but the WX project has not added a way to
do that for these types of macro deprecations. So we add a way to
re-define the wx deprecations macro to not emit any warning.