compiler: beam_ssa_ss: Enforce cutoff for list literals
The alias analysis pass tracks the structure and aliasing status of
values given as arguments and returned from functions. The information
is kept in the graph provided by the beam_ssa_ss module and is built
when the alias analysis processes instructions constructing
terms. When a literal is given as an argument to a function, the graph
is extended to have the same structure as if the literal was
constructed programmatically.

To handle self-recursive functions which, for example builds a list,
there is a cutoff which limits the depth to which the graph is
constructed. This allows a fixpoint to be reached without implementing
a solver for recurrence equations about the structure of the
arguments, but also limits the size of the graph.

This patch corrects an omission in the beam_ssa_ss:merge_in_arg/4
function where the cutoff was not enforced for list literals.

Closes #8140