stdlib: Add warn_missing_spec_documented
With the addition of -doc attributes, it often is desired to have a spec
on those functions. This warning makes sure that the code will always
have that.

It was considered to just change the behaviour of `warn_missing_spec`,
but we decided to go the conservative route as any code that today has
`warn_missing_spec` and decided to use -doc attributes would also need
to remember to update the compiler flag.