Correct coverage for functions on the same line
This commit fixes the bug reported in #8159.

It also eliminates a weird behavior when there are multiple functions
on the same line. Consider this module:

    -module(foo).
    -export([bar/0, baz/0]).

    bar() -> ok. baz() -> not_ok.

In Erlang/OTP 26, analysing on the line level would return two entries
for line 4:

    1> cover:compile_module(foo).
    {ok,foo}
    2> foo:bar().
    ok
    3> cover:analyse(foo, coverage, line).
    {ok,[{{foo,4},{1,0}},{{foo,4},{0,1}}]}
    4> cover:analyse(foo, calls, line).
    {ok,[{{foo,4},1},{{foo,4},0}]}

This has been changed to coalesce all information about each line to
a single entry in the resulting list:

    1> cover:compile_module(foo).
    {ok,foo}
    2> foo:bar().
    ok
    3> cover:analyse(foo, coverage, line).
    {ok,[{{foo,4},{1,0}}]}
    4> cover:analyse(foo, calls, line).
    {ok,[{{foo,4},1}]}

Closes #8159