compiler: beam_ssa_alias: Improve analysis and make it deterministic
The best, i.e. least conservative, alias analysis result appears to be
reached when a function's callers are analyzed before the function
itself, i.e. what is needed is a reverse post-order traversal of the
call graph.

This patch corrects an error where exported functions with local
callers could be analyzed before exported functions without local
callers. It also makes the order in which functions are analyzed
deterministic.