erts: Fix erts_code_purger to not drop finish_after_on_load messages
while doing a purge operation in cpc_receive().

The symptom will be code_server process hanging in
erts_code_purger:finish_after_on_load/2 waiting for reply.

This is probably not a problem as long as only the code server
interface (code) is used to load and purge code and not the
discouraged lower level functions like erlang:purge_module.