compiler: Do not emit doc chunk if no -doc features are used
If a module does not have any -doc attributes, it is not clear
if we use EEP-59 or not in the system. So for edoc_*_chunks to work
properly we should not emit any doc chunk as otherwise anyone using edoc
will always have to pass no_docs to the compiler.