Eliminate redundant test for empty binary
Consider this module:

    -module(test).
    -export([test/1]).

    test(Bin) when is_binary(Bin) ->
       test1(Bin).

    test1(<<_, Rest/binary>>) -> test1(Rest);
    test1(<<>>) -> ok.

The only way the first clause of `test1/1` can fail is when the input
binary is empty. Therefore, testing for an empty binary in the second
clause is redundant. The compiler still generates a test for
an empty binary and code to raise a `function_clause` exception if the
binary is not empty. Because an exception can be raised, the compiler
also needs to insert a book-keeping instruction (`bs_set_position`)
into the code for the first clause. That extra instruction has a
measurable impact on performance.

This commit eliminates the test for an empty binary and the extra
`bs_set_position` instruction when it is safe to do so.

Thanks to Michał Muskała for noticing this issue.