compiler: Avoid infinite loops in destructive update pass
The destructive update pass can, with source code of a particular
structure, enter an infinite loop when trying to find initial values
which need to be moved to the heap. The infinite loop is caused by
initial value tracker exploring a non-terminating chain of
`get_tuple_element` or `get_hd` instructions.

As code triggering the infinite loop doesn't appear to very common,
just impose a limit on the depth to which the search for initial
values is allowed to reach. If the limit is reached, destructive
updates are disabled for the module.

Note that two test modules are added, this is to get coverage of both
error paths in the compiler, as the first detected case stops
processing for the entire module.

Closes #8281