compiler: Do strength reduction of the hint for update_record
Change the hint from `reuse` to `copy` when reusing clearly
will not work.

Extending this optimization to be module-global, and not just work at
the function level, increases the time spent on this pass by a factor
of between 4 and 5 and only triggers infrequently, and for example,
never allows for additional destructive updates. For the moment this
is considered good enough.

Co-authored-by: Frej Drejhammar <frej.drejhammar@gmail.com>