Fix del_table_copy loading loop
When a local table copy was deleted but not yet loaded mnesia could
try to load it in a loop, which later when a table copy was added
again could result in a crash, since we don't have locks on the
receiver anymore. Tables can be simultaneously loaded and added.

Fixed by removing the table from the loader queue when it is
determined that the local copy is removed.

Also don't 'fatal' crash mnesia if table couldn't be loaded due
to table already existing.