archives: Formally deprecate and schedule for removal
Archives is an experimental feature. Its mere presence makes
its hard or impossible to improve the time for starting `erl`.

Warn for the removal of archives for the purpose of holding a single
application and including those archives in the code path.

Using a single archive in an Escript is **not** deprecated.