tprof: Fix profiling output to not include tprof code
In order to make sure that tprof does not include itself in the
profiling output when profiling we make sure to reset the counters
just before we start the profiling and pause them just after we care done.

Also we filter the calls to the function used to pause the counters
so that we only get what was actually profiled.

Timeout handling has also been reworked in in order to not trigger
any code that will be seen as part of what is profiled.