asn1: Retain the definition order in the JER encoding
Consider this ASN.1 module:

    TT DEFINITIONS AUTOMATIC TAGS ::=
    BEGIN
       S1 ::= SEQUENCE {
         nisse INTEGER,
         arne INTEGER,
         kalle INTEGER,
         fredrik INTEGER
       }
    END

When encoding the `S1` SEQUENCE to JER (JSON encoding rules) using one
of the pre-releases of Erlang/OTP 27 that includes the new `json`
module, the order of the keys and value in the definition of `S1`
would not be retained:

    1> asn1ct:compile('TT', [jer]).
    ok
    2> io:format("~ts\n", [element(2, 'TT':encode('S1', {'S1',1,2,3,4}))]).
    {"arne":2,"fredrik":4,"kalle":3,"nisse":1}
    ok

While technically not a bug, because the order is not documented, it
is a change from how Erlang/OTP 26 and earlier did the JSON encoding,
and some use cases that depended on that order could break.
Also, it can be helpful for debugging to keep the order.

Therefore, this commit ensures that the order is retained:

    1> asn1ct:compile('TT', [jer]).
    ok
    2> io:format("~ts\n", [element(2, 'TT':encode('S1', {'S1',1,2,3,4}))]).
    {"nisse":1,"arne":2,"kalle":3,"fredrik":4}
    ok

Closes #8453