Get rid of macros for bit syntax construction
The first implementation of the bit syntax was done in Erlang/OTP R7B
before the SMP emulator (the multi-threaded runtime system).
Therefore, global variables were used to keep track of the binary
being constructed.

When the SMP emulator was introduced, the state of the
binary being constructed needed to be local for each
scheduler thread. To reduce the need for extensive rewrites,
macros were used to automagically pass references
to the binary construction state to the helper functions
in erl_bits.c.

Now bite the bullet and remove these macros. Also take the opportunity
to remove the "new" part from function names. They are not exactly
new.