stdlib: turn singleton type var warning into error
Singleton type variables in an union type do not make sense from
Dialyzer's point of view. The following example is ill-typed:

```erlang
-spec run_test(Opts) -> term()
      when Opts :: {join_specs, Bool} | {test, Bool}.
```

But this cannot be ever checked by Dialyzer as the type variable `Bool`
is in a union and used a single time. Dialyzer cannot check that the
`Bool` set of values passed to `join_specs` will be the same set of
values as the ones passed in the `Bool` type variable in the `test`
branch. Thus, an error follows.