[esock|test] Add wrapper test module for running all socket suites
The kernel app has *a lot* of (test) suites so running
the entire kernal app (test) is a bit much, and running
all the socket suites individually is a bit annoying.
So, added a wrapper file, socket_suites, that you run as
a suite and that in turn runs "all" the socket suites.