compiler: Limit depth of sharing state tree
With OTP 27.0 the xmerl tests showed a markedly increased run-time.
Profiling pointed to beam_ssa_ss:prune/2 as the culprit. It turns out
that if there are long value derivation chains which are extended one
element at a time (such as in the ss_depth_limit.erl test-case in this
patch), the effective time complexity of prune will be O(n^2) where n
is the number of elements in the chain.

By limiting the depth of the sharing state tree, we can avoid the
drastically increased run-times for the alias analysis. This is a
stop-gap fix until the alias analysis can be improved to avoid this
pit-fall.