Refactor type annotation handling for clarity
In #8564, Frej Drejhammar noticed that the following comment
was misleading:

    %% One or more arguments have been simplified to literal values.

It turns out that the entire clause in which the comment is located is
not needed. Removing the `arg_types` annotation is unnecessary, since
it will be added back in the next iteration of the `beam_ssa_type`
pass.

This commit removes the unnecessary clause, as well as strengthening
and explaining the invariants maintained by `opt_anno_types/2`.