Implement native records
This is an implementation of native records as described in
EEP-79 (https://github.com/erlang/eep/pull/81).

Note that EEP-79 has not been finalized. Therefore, there will be
updates to the EEP and possibly also to the implementation before the
second release candidate. Notable discrepancies between the EEP and
the implementation is type definitions for native records. The
parameterization of record types is not supported in the current
implementation.

Support for native records has not yet been implemented in the
`debugger`, `erl_interface`, and `jinterface` applications.
Also, there is currently no support for native records in NIFs.

There is partial support for native records in Dialyzer. Currently,
types for record elements are not checked.

Co-authored-by: John Högberg <john@erlang.org>
Co-authored-by: Isabell Huang <isabell@erlang.org>
Co-authored-by: Ilya Klyuchnikov <ilya.klyuchnikov@gmail.com>
Co-authored-by: Dan Gudmundsson <dgud@erlang.org>