ssl: Add signature_algs_cert default when signature_algs default is used
Make upgrade path smoother by adding rsa_pkcs1_sha to the
default of signature_algs as the default signature_algs_cert.
Note this is only applicable when signature_algs is not configured,
that is set to the default, that will then become the default of
signature_algs_cert in practice. This will allow certificates to
use rsa_pkcs1_sha algorithm but still disallow it in the TLS protocol.

Also add some missing handling of signature_algs_cert in DTLS.

closes #8588