[erts] Fix monotonicity of Erlang monotonic time
When 'no time warp mode' was enabled, a smaller Erlang monotonic time
could be read than a previously read time, i.e., breaking the monotonic
property. The runtime system will abort when detecting an issue like
this since OTP 24.3.4.17 and OTP 25.0.

On systems with a fine grained time unit, the window for this happening
was very small, but still existed. On systems with native nanosecond
time unit, which are most if not all non-Windows systems, two readings
needed to be done within 20 nanoseconds.