compiler alias analysis: Avoid redundant database changes
This is a micro optimization which avoids creating a variable in the
sharing database when, during the processing of the instruction, we
will immediately update its status. This reduces the number of sharing
database operations as we can insert and set the status in a single
operation. It also allows us to completely avoid creating a variable
for extracted plain values.