compiler: Fix crash in beam_ssa_private_append
With the right input, the beam_ssa_private_append pass could crash
when it, during the initial value tracking phase, selected literals
which were not compatible with the tracked element structure for later
patching. This patch ensures that only compatible literals are
scheduled for later patching.

This crash is only present in OTP versions >= 26 && <= 27 and this fix
should not be merged to master as dce585f933d0 ("compiler: In-place
update of tuples/records"), merged for OTP 27, includes essentially
the same functionality.

Closes #8630