compiler tests: Correct bad timetrap limit
The compiler's random_code test suite always fails when running with a
debug emulator due to a `timeout_value` exit. A `timeout_value` exit
appears to be triggered when a timer is given a timeout that is either
miss-formed or out of range. Checking the relevant part of
`random_code_SUITE`, it appears that the timeout is calculated in
milliseconds, but it is given to `ct:timetrap/1` as a timeout in
seconds.

Although not sure about the exact failure mechanism, changing the
calculation to produce a timeout in seconds, allows the test-suite to
successfully complete both with an `opt` and `debug` build of the
emulator.