compiler alias analysis: Improve fixpoint iteration strategy
Switch from a dumb fixpoint iteration where we just iterated until the
sharing state for a all functions stopped changing, to a more
intelligent strategy where we only consider changes that are visible
outside the function. Concretely that means that we only re-analyze
callers if the callee's result's status changed. And callees if
argument aliasing status has changed.

The new strategy leads to a test failure (marked xfail), but that
limitiation will be removed in a later patch.