compiler alias analysis: Introduce explicit no_info alias status
Previously the status of a variable in the sharing state database was
either `unique` or `aliased`. This patch extends the set of statuses
with a third: `no_info`, meaning that nothing is known about the alias
status of the variable, nor of its structure. This new status will be
used in future changes in order to allow for faster and more efficient
calculation of the alias status of a module.