compiler alias analysis: Speed up state pruning
In around 80% of the cases when prune is called, more than half of the
nodes in the sharing state database survive. Therefore a pruning
strategy which removes nodes from the database has been used. This
patch resurrects the previously used pruning algorithm which instead
recreated the pruned state from scratch instead of removing nodes
which is faster when only a few nodes survive the pruning. The alias
analysis implementation is then extended to, for each pruned basic
block, by looking at the number of nodes in the database before and
after the prune, record which strategy is the most efficient. If the
basic block is revisited, the fastest prune implementation is
selected. On the modules compiled by the `scripts/diffable` tool, this
change reduces the time spent on alias analysis by close to six
percent.