compiler: Add support for debug information in BEAM files
The `beam_debug_info` compiler option will insert `debug_line`
instructions on lines containing executable code, and it will maintain
information about which variables the BEAM registers contain at each
`debug_line` instruction. This information will be inserted into a
"DbgB" chunk in the BEAM file.

When a `debug_line` is executed, the current stack frame (if any) is
guaranteed to be fully initialized. The number of live X registers is
given as the second operand for the `debug_line` instruction (it is
guaranteed that there are no "holes").

Here is an example where the debug information translated to text has
been inserted as comments before the lines they apply to:

    %% function entry (no stack frame); x0, x1, x2 are live
    sum(A, B, _Ignored) ->
	%% no stack frame; A in x0, B in x1, _Ignored in x2
	C = A + B,

	%% no stack frame; B in x1, C in x0
	io:format("~p\n", [C]),

	%% stack frame size is 1; C in y0
	D = 10 * C,

	%% stack frame size is 1; C in y0, D in x0
	{ok,D}.

Note that not all variables are available in the debug information.

For example, before the call to `io:format/2`, the sum of A and B have
overwritten the register that used to hold the value of A, and the value
for _Ignore was wiped out by the `+` operation.

The size of the current stack frame is also given at each `debug_line`
to be able to easily find the beginning of the previous stack frame.