compiler destructive update: Patch every element of a list
The initial value tracker in the destructive update pass could
sometimes fail to ensure that a term which was destructively updated
was forced onto the heap as it did not have a way to easily follow
list tails. This patch corrects that error by ensuring that if the
first element of a literal list is forced onto the heap, all
elements of the list will be forced onto the heap.